+++ bench/type4/coverage_probes/import_identity/swift/neg-member/b.swift
@@ -1,0 +1,5 @@
+import Shared
+
+func axisCase(_ value: Int) -> Int {
+    return Shared.other(value + 1)
+}