+++ bench/type4/coverage_probes/import_identity/swift/neg-module/b.swift
@@ -1,0 +1,5 @@
+import Other
+
+func axisCase(_ value: Int) -> Int {
+    return Other.helper(value + 1)
+}