+++ bench/type4/coverage_probes/total_order_compare/swift/neg-domain/a.swift
@@ -1,0 +1,3 @@
+func axisCase(_ x: String, _ y: String) -> Bool {
+    return x < y && x <= y
+}