+++ bench/type4/coverage_probes/total_order_compare/swift/neg-operator/a.swift
@@ -1,0 +1,3 @@
+func axisCase(_ x: Int, _ y: Int) -> Bool {
+    return x < y || x <= y
+}